convert file to typescript
